Refactor: move core module to src folder
