+++ packages/core/src/resources/console.ts
