+++ packages/core/src/resources/resource.ts
