+++ packages/core/src/tools/snapshot.ts
