+++ packages/core/src/tools/tool.ts
